QPID-2996 - qmf/SchemaMethod.cpp fails to compile on s390 Applied patch from Dan Horák
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid@1060815 13f79535-47bb-0310-9956-ffa450edef68

+++ cpp/src/qmf/SchemaMethod.cpp
@@ -163,7 +163,7 @@ void SchemaMethodImpl::encodeV1(qpid::management::Buffer& buffer) const
     Variant::Map map;
 
     map["name"] = name;
-    map["argCount"] = arguments.size();
+    map["argCount"] = (uint64_t) arguments.size();
     if (!desc.empty())
         map["desc"] = desc;
